added host options, bugfixes and improvments
